54120 funcs

+++ include/functions.us.h
@@ -432,7 +432,7 @@ void* MallocPerm(s32 size);
 s16 func_80060288(void);
 void func_80018C90(u16);
 s32 func_80019060(s32, s16, s16);
-void func_80067480(s16, s32, s32); //could also be  void func_80067480(s16, s32, u16); (as required by functions in face lift)
+void func_80067480(s16, s16, s32); //could also be  void func_80067480(s16, s32, u16); (as required by functions in face lift)
 void func_800672DC(s16, s16, u16, s32);
 void func_800674BC(s16, s16, u16);
 s16 func_80064EF4(s32, s32);
